requirements/default.txt
#pdf
pypdf==5.3.0
PyMuPDF>=1.25.3
pycryptodome==3.21.0

# markdown to pdf
weasyprint==64.1
markdown==3.7

# docx
python-docx==1.1.2

# google generative ai
google-genai==1.14.0

# audio video processing
ffmpeg-python==0.2.0